Add plutus-emulator package
