+++ lib.nix
@@ -39,6 +39,7 @@ let
     "marlowe"
     "meadow"
     "plutus-wallet-api"
+    "plutus-emulator"
   ];
 
   plutusHaskellPkgList = lib.filter (v: v != "plutus-playground-client" && v != "plutus-server-invoker") plutusPkgList;
