+++ plutus-emulator/README.md
@@ -1,0 +1,3 @@
+# plutus-emulator
+
+Implementation of an extended UTXO ledger with scripts (the mockchain).
