+++ plutus-emulator/shell.nix
@@ -1,0 +1,1 @@
+(import ../. {}).shellTemplate "plutus-emulator"
