+++ plutus-emulator/src/Wallet/Emulator.hs
