+++ plutus-playground-lib/plutus-playground-lib.cabal
@@ -51,6 +51,7 @@ library
     , text
     , transformers
     , plutus-wallet-api
+    , plutus-emulator
   default-language: Haskell2010
 
 test-suite plutus-playground-lib-test
@@ -75,6 +76,7 @@ test-suite plutus-playground-lib-test
         template-haskell -any,
         plutus-playground-lib -any,
         plutus-wallet-api -any,
+        plutus-emulator -any,
         hspec -any,
         QuickCheck -any,
         aeson -any
