+++ plutus-use-cases/test/Spec/Game.hs
@@ -16,6 +16,7 @@ import qualified Ledger
 import qualified Ledger.Ada                                    as Ada
 import qualified Ledger.Value                                  as Value
 import           Wallet.Emulator
+import qualified Wallet.Emulator.Generators                    as Gen
 import qualified Wallet.Generators                             as Gen
 
 import           Language.PlutusTx.Coordination.Contracts.Game (gameValidator, guess, lock, startGame)
