+++ plutus-wallet-api/src/Wallet/Generators.hs
@@ -28,8 +28,6 @@ module Wallet.Generators(
     genValueNonNegative,
     genSizedByteString,
     genSizedByteStringExact,
-    Wallet.Generators.runTrace,
-    runTraceOn,
     splitVal,
     validateMockchain,
     signAll
@@ -55,7 +53,6 @@ import qualified Ledger.Value    as Value
 
 import           Ledger
 import qualified Wallet.API      as W
-import           Wallet.Emulator as Emulator
 
 -- | Attach signatures of all known private keys to a transaction.
 signAll :: Tx -> Tx
@@ -262,20 +259,6 @@ validateMockchain (Mockchain blck _) tx = either Just (const Nothing) result whe
     h      = lastSlot [blck]
     idx    = Index.initialise [blck]
     result = Index.runValidation (Index.validateTransaction h tx) idx
-
--- | Run an emulator trace on a mockchain.
-runTrace ::
-    Mockchain
-    -> Trace MockWallet a
-    -> (Either AssertionError a, EmulatorState)
-runTrace (Mockchain ch _) = Emulator.runTraceTxPool ch
-
--- | Run an emulator trace on a mockchain generated by the model.
-runTraceOn :: MonadGen m
-    => GeneratorModel
-    -> Trace MockWallet a
-    -> m (Either AssertionError a, EmulatorState)
-runTraceOn gm t = flip Wallet.Generators.runTrace t <$> genMockchain' gm
 
 -- | Split a value into max. n positive-valued parts such that the sum of the
 --   parts equals the original value.
